Organise swap-related confirmation views entities
