+++ src/routes/transactions/transactions-view.controller.ts
@@ -10,8 +10,9 @@ import {
 import {
   BaselineConfirmationView,
   ConfirmationView,
-  CowSwapConfirmationView,
 } from '@/routes/transactions/entities/confirmation-view/confirmation-view.entity';
+import { CowSwapConfirmationView } from '@/routes/transactions/entities/swaps/swap-confirmation-view.entity';
+import { CowSwapTwapConfirmationView } from '@/routes/transactions/entities/swaps/twap-confirmation-view.entity';
 import { NativeStakingDepositConfirmationView } from '@/routes/transactions/entities/staking/native-staking-deposit-confirmation-view.entity';
 import { NativeStakingValidatorsExitConfirmationView } from '@/routes/transactions/entities/staking/native-staking-validators-exit-confirmation-view.entity';
 import { NativeStakingWithdrawConfirmationView } from '@/routes/transactions/entities/staking/native-staking-withdraw-confirmation-view.entity';
@@ -54,6 +55,7 @@ export class TransactionsViewController {
       oneOf: [
         { $ref: getSchemaPath(BaselineConfirmationView) },
         { $ref: getSchemaPath(CowSwapConfirmationView) },
+        { $ref: getSchemaPath(CowSwapTwapConfirmationView) },
         { $ref: getSchemaPath(NativeStakingDepositConfirmationView) },
         { $ref: getSchemaPath(NativeStakingValidatorsExitConfirmationView) },
         { $ref: getSchemaPath(NativeStakingWithdrawConfirmationView) },
@@ -63,6 +65,8 @@ export class TransactionsViewController {
   @ApiExtraModels(
     BaselineConfirmationView,
     CowSwapConfirmationView,
+    // Prevent bidirectional dependency
+    () => CowSwapTwapConfirmationView,
     NativeStakingDepositConfirmationView,
     NativeStakingValidatorsExitConfirmationView,
     NativeStakingWithdrawConfirmationView,
